% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{button_extra}
\alias{button_extra}
\title{Button for reactable}
\usage{
button_extra(id, ...)
}
\arguments{
\item{id}{id of the button input}

\item{...}{parameters of button}
}
\description{
Button for reactable
}
