% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{checkbox_extra}
\alias{checkbox_extra}
\title{Checkbox input for reactable column cell}
\usage{
checkbox_extra(id, ...)
}
\arguments{
\item{id}{id of the checkbox input}

\item{...}{parameters of checkbox, only `class` is supported for now}
}
\description{
Checkbox input for reactable column cell
}
\examples{
colDef(cell = checkbox_extra("check", class = "table-check"))

}
