% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{date_extra}
\alias{date_extra}
\title{Date input for reactable column cell}
\usage{
date_extra(id, ...)
}
\arguments{
\item{id}{id of the date input}

\item{...}{parameters of date input, only `class` is supported for now}
}
\description{
Date input for reactable column cell
}
\examples{
reactable::colDef(cell = date_extra("date", class = "table-date"))

}
