% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{define_key}
\alias{define_key}
\title{Define the unique id to use when passing values to shiny}
\usage{
define_key(key)
}
\description{
Define the unique id to use when passing values to shiny
}
\keyword{internal}
