% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable-server.R
\name{get_data_on_page}
\alias{get_data_on_page}
\title{Get the data on the page}
\usage{
get_data_on_page(data, page_number, total_pages)
}
\arguments{
\item{data}{data.frame, an unpaged data; pagination will be inferred}

\item{page_number}{page of data to retrieve}

\item{total_pages}{number of pages in \code{data}; determines pagination when the data has no pages}
}
\value{
a data.frame
}
\description{
Get the data on the page
}
\keyword{internal}
