% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable-server.R
\name{reactable-page-controls}
\alias{reactable-page-controls}
\alias{reactable_page_controls}
\alias{return_reactable_page}
\title{Module for reactable page navigation}
\usage{
reactable_page_controls(id)

return_reactable_page(id, total_pages)
}
\arguments{
\item{id}{element id}

\item{total_pages}{total number of pages}
}
\value{
\code{reactable_page_controls()} returns a UI for page navigation of a server-side processed
\code{\link[reactable:reactable]{reactable::reactable()}} data
}
\description{
Module for reactable page navigation
}
\keyword{internal}
