% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{reactable_extras_dependency}
\alias{reactable_extras_dependency}
\title{Reactable.extras JS and CSS dependencies}
\usage{
reactable_extras_dependency()
}
\value{
HTML dependency object.
}
\description{
Reactable.extras JS and CSS dependencies
}
