% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{reactableExtrasDependency}
\alias{reactableExtrasDependency}
\title{htmlDependency of reactable server including script and styles}
\usage{
reactableExtrasDependency()
}
\value{
An htmlDependency object
}
\description{
htmlDependency of reactable server including script and styles
}
