% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{text_extra}
\alias{text_extra}
\title{Text input for reactable column cell}
\usage{
text_extra(id, key = NULL, ...)
}
\arguments{
\item{id}{id of the text input}

\item{key}{alternative unique id for server side processing}

\item{...}{parameters of text input, only \code{class} is supported for now}
}
\value{
Custom JS text input renderer for reactable
}
\description{
Text input for reactable column cell
}
\examples{
reactable::colDef(cell = text_extra("text", class = "table-text"))

}
