% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable-server.R
\name{toggle_navigation_buttons}
\alias{toggle_navigation_buttons}
\title{Utility function to disable or re-enable navigation buttons}
\usage{
toggle_navigation_buttons(disable, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{disable}{a named logical vector}

\item{session}{Shiny session object; default to current Shiny session}
}
\description{
Utility function to disable or re-enable navigation buttons
}
\details{
\code{disable} should a logical vector with these exact names: \code{first_page}, \code{previous_page},
\code{next_page}, and \code{last_page}. The logical vectors indicate if the corresponding button will be
enabled or disabled.
}
\keyword{internal}
