% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooltip.R
\name{tooltip_extra}
\alias{tooltip_extra}
\title{Tool-tip for table headers}
\usage{
tooltip_extra(content, theme = "light")
}
\arguments{
\item{content}{The content to be displayed in the tool-tip}

\item{theme}{The theme of the tool-tip, either "light", "light-border", "material" or
"translucent"}
}
\value{
Custom JS tool-tip renderer for reactable
}
\description{
Tool-tip for table headers
}
\examples{
reactable::colDef(header = tooltip_extra("This is my tool-tip", theme = "material"))

}
