% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{button_extra}
\alias{button_extra}
\title{Button input for reactable column cell}
\usage{
button_extra(id, ...)
}
\arguments{
\item{id}{id of the button input}

\item{...}{parameters of button, only `class` is supported for now}
}
\description{
Button input for reactable column cell
}
\examples{
colDef(cell = button_extra("click", class = "table-button"))


}
