% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{dropdown_extra}
\alias{dropdown_extra}
\title{Select input for reactable's column cell}
\usage{
dropdown_extra(id, choices, ...)
}
\arguments{
\item{id}{id of the select input}

\item{choices}{vector of choices}

\item{...}{parameters of date input, only `class` is supported for now}
}
\description{
Select input for reactable's column cell
}
\examples{
colDef(cell = dropdown_extra("dropdown", class = "table-dropdown"))

}
