% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{dropdown_extra}
\alias{dropdown_extra}
\title{Select input for reactable column cell}
\usage{
dropdown_extra(id, choices, key = NULL, ...)
}
\arguments{
\item{id}{id of the select input}

\item{choices}{vector of choices}

\item{key}{alternative unique id for server side processing}

\item{...}{parameters of date input, only \code{class} is supported for now}
}
\value{
Custom JS dropdown renderer for reactable
}
\description{
Select input for reactable column cell
}
\examples{
reactable::colDef(
  cell = dropdown_extra("dropdown",
  choices = letters[1:5],
  class = "table-dropdown"
  )
 )

}
