% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable-server.R
\name{reactable-extras-server}
\alias{reactable-extras-server}
\alias{reactableExtrasUi}
\alias{reactableExtrasServer}
\title{Create reactable UI with server-side processing}
\usage{
reactableExtrasUi(id, width = "auto", height = "auto")

reactableExtrasServer(id, data, rows_per_page = 10, sortable = TRUE, ...)
}
\arguments{
\item{id}{element id}

\item{width, height}{CSS unit (\code{"100\%"}, \code{"400px"}, or \code{"auto"}), numeric for number of pixels}

\item{data}{passed to \code{\link[reactable:reactable]{reactable::reactable()}}}

\item{rows_per_page}{number of pages to show}

\item{sortable}{allow sorting by columns}

\item{...}{other arguments to be passed to \code{\link[reactable:reactable]{reactable::reactable()}}}
}
\value{
\code{reactableExtrasUi()} returns a custom UI for a server-side processed reactable
}
\description{
Create reactable UI with server-side processing
}
