% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{reactable_extras_dependency}
\alias{reactable_extras_dependency}
\title{htmlDependency of reactable server including script and styles}
\usage{
reactable_extras_dependency()

reactable_extras_dependency()
}
\value{
An htmlDependency object

HTML dependency object.
}
\description{
htmlDependency of reactable server including script and styles

Shiny Fluent JS dependency
}
