% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{reactableExtrasDependency}
\alias{reactableExtrasDependency}
\title{Shiny Fluent JS dependency}
\usage{
reactableExtrasDependency()
}
\value{
HTML dependency object.
}
\description{
Shiny Fluent JS dependency
}
