% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tooltip.R
\name{tooltip_extra}
\alias{tooltip_extra}
\title{Tooltip for table headers}
\usage{
tooltip_extra(content, theme = "light")
}
\arguments{
\item{content}{The content to be displayed in the tooltip}

\item{theme}{The theme of the tooltip, either "light", "light-border", "material" or
"translucent"}
}
\description{
Tooltip for table headers
}
\examples{
reactable::colDef(header = tooltip_extra("This is my tooltip", theme = "material"))

}
